Implement things as "related items"
We cannot escape the fact that ItemSummaryComponent needs some knowledge of what's going on in order to display links and a modal. We can pretend this is not a layer violation by trying to generalize a little bit. We keep the display property in the camera service as plain text, should we need it in the future, but the rest happens in the summary item.